requirements.txt
gymnasium>=0.29
numpy>=1.24
pyyaml>=6.0
stable-baselines3>=2.0
PyQt6>=6.5
matplotlib>=3.7
